add library to access freebase.
adapted from https://github.com/tombenner/freeb/, main changes are in topic.rb and API.rb. The changes are to make this work as standalone (not as gem) and to with our version of ruby. Big feature we are missing right now is the ability to cache results.

note: only the search function is tested. currently only using this function.